moved load account history to use unlocked account, fixed unit tests.
