+++ src/modules/auth/actions/set-login-account.js
@@ -1,5 +1,4 @@
 import { useUnlockedAccount } from 'modules/auth/actions/use-unlocked-account'
-import { loadAccountHistory } from 'modules/auth/actions/load-account-history'
 import logError from 'utils/log-error'
 
 // If there is an available unlocked account, set as the user's sending address.
@@ -7,6 +6,5 @@ export const setLoginAccount = (autoLogin, account, callback = logError) => (dis
   // If autoLogin=true, use an unlocked local Ethereum node (if present)
   if (autoLogin) {
     dispatch(useUnlockedAccount(account, callback))
-    dispatch(loadAccountHistory())
   }
 }
