+++ src/modules/my-positions/selectors/my-positions.js
@@ -11,8 +11,10 @@ export const selectPositionsMarkets = createSelector(
   selectMarkets,
   selectAccountPositionsState,
   (markets, positions) => {
-    if (!markets || markets.length === 0) return []
-    if (!positions || positions.length === 0) return []
-    return (markets || []).filter(market => positions[market.id] != null)
+    if (!markets || !positions) {
+      return []
+    }
+    (markets || []).filter(market => Object.keys(positions || {})
+      .find(positionMarketID => market.id === positionMarketID))
   }
 )
