+++ test/auth/actions/set-login-account-test.js
@@ -21,14 +21,11 @@ describe(`modules/auth/actions/set-login-account.js`, () => {
         }
       }
       const UseUnlockedAccount = { useUnlockedAccount: () => {} }
-      const LoadAccountHistory = { loadAccountHistory: () => {} }
       const action = proxyquire('../../../src/modules/auth/actions/set-login-account.js', {
         '../../../services/augurjs': AugurJS,
-        './load-account-history': LoadAccountHistory,
         './use-unlocked-account': UseUnlockedAccount
       })
       sinon.stub(UseUnlockedAccount, 'useUnlockedAccount').callsFake(account => ({ type: 'USE_UNLOCKED_ACCOUNT', account }))
-      sinon.stub(LoadAccountHistory, 'loadAccountHistory').callsFake(account => ({ type: 'LOAD_ACCOUNT_HISTORY', data: {} }))
       store.dispatch(action.setLoginAccount(t.params.autoLogin, t.params.account))
       t.assertions(store.getActions())
       store.clearActions()
@@ -112,10 +109,6 @@ describe(`modules/auth/actions/set-login-account.js`, () => {
       assert.deepEqual(actions, [{
         type: 'USE_UNLOCKED_ACCOUNT',
         account: '0xtest'
-      },
-      {
-        type: 'LOAD_ACCOUNT_HISTORY',
-        data: {}
       }])
     }
   })
